+++ src/main/kotlin/app/revanced/api/configuration/repository/ConfigurationRepository.kt
@@ -1,7 +1,9 @@
 package app.revanced.api.configuration.repository
 
+import app.revanced.api.configuration.schema.APIAbout
 import app.revanced.api.configuration.services.ManagerService
 import app.revanced.api.configuration.services.PatchesService
+import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -10,7 +12,11 @@ import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
 import kotlinx.serialization.descriptors.SerialDescriptor
 import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
+import kotlinx.serialization.json.Json
+import kotlinx.serialization.json.JsonNamingStrategy
+import kotlinx.serialization.json.decodeFromStream
 import java.io.File
+import java.nio.file.Path
 
 /**
  * The repository storing the configuration for the API.
@@ -24,6 +30,10 @@ import java.io.File
  * @property corsAllowedHosts The hosts allowed to make requests to the API.
  * @property endpoint The endpoint of the API.
  * @property oldApiEndpoint The endpoint of the old API to proxy requests to.
+ * @property staticFilesPath The path to the static files to be served under the root path.
+ * @property versionedStaticFilesPath The path to the static files to be served under a versioned path.
+ * @property about The path to the json file deserialized to [APIAbout]
+ * (because com.akuleshov7.ktoml.Toml does not support nested tables).
  */
 @Serializable
 internal class ConfigurationRepository(
@@ -40,6 +50,15 @@ internal class ConfigurationRepository(
     val endpoint: String,
     @SerialName("old-api-endpoint")
     val oldApiEndpoint: String,
+    @Serializable(with = PathSerializer::class)
+    @SerialName("static-files-path")
+    val staticFilesPath: Path,
+    @Serializable(with = PathSerializer::class)
+    @SerialName("versioned-static-files-path")
+    val versionedStaticFilesPath: Path,
+    @Serializable(with = AboutSerializer::class)
+    @SerialName("about-json-file-path")
+    val about: APIAbout,
 ) {
     /**
      * Am asset configuration whose asset is signed.
@@ -107,4 +126,24 @@ private object FileSerializer : KSerializer<File> {
     override fun serialize(encoder: Encoder, value: File) = encoder.encodeString(value.path)
 
     override fun deserialize(decoder: Decoder) = File(decoder.decodeString())
+}
+
+private object PathSerializer : KSerializer<Path> {
+    override val descriptor: SerialDescriptor = PrimitiveSerialDescriptor("Path", PrimitiveKind.STRING)
+
+    override fun serialize(encoder: Encoder, value: Path) = encoder.encodeString(value.toString())
+
+    override fun deserialize(decoder: Decoder): Path = Path.of(decoder.decodeString())
+}
+
+private object AboutSerializer : KSerializer<APIAbout> {
+    override val descriptor: SerialDescriptor = PrimitiveSerialDescriptor("APIAbout", PrimitiveKind.STRING)
+
+    override fun serialize(encoder: Encoder, value: APIAbout) = error("Serializing APIAbout is not supported")
+
+    @OptIn(ExperimentalSerializationApi::class)
+    val json = Json { namingStrategy = JsonNamingStrategy.SnakeCase }
+
+    override fun deserialize(decoder: Decoder): APIAbout =
+        json.decodeFromStream(File(decoder.decodeString()).inputStream())
 }