+++ src/main/kotlin/app/revanced/api/configuration/routes/ApiRoute.kt
@@ -1,9 +1,11 @@
 package app.revanced.api.configuration.routes
 
+import app.revanced.api.configuration.*
 import app.revanced.api.configuration.installCache
 import app.revanced.api.configuration.installNoCache
 import app.revanced.api.configuration.installNotarizedRoute
 import app.revanced.api.configuration.respondOrNotFound
+import app.revanced.api.configuration.schema.APIAbout
 import app.revanced.api.configuration.schema.APIContributable
 import app.revanced.api.configuration.schema.APIMember
 import app.revanced.api.configuration.schema.APIRateLimit
@@ -13,7 +15,6 @@ import io.bkbn.kompendium.core.metadata.*
 import io.ktor.http.*
 import io.ktor.server.application.*
 import io.ktor.server.auth.*
-import io.ktor.server.http.content.*
 import io.ktor.server.plugins.ratelimit.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
@@ -56,6 +57,16 @@ internal fun Route.apiRoute() {
         }
     }
 
+    route("about") {
+        installCache(1.days)
+
+        installAboutRouteDocumentation()
+
+        get {
+            call.respond(apiService.about)
+        }
+    }
+
     route("ping") {
         installNoCache()
 
@@ -75,9 +86,21 @@ internal fun Route.apiRoute() {
             }
         }
 
-        staticResources("/", "/app/revanced/api/static/versioned") {
-            contentType { ContentType.Application.Json }
-            extensions("json")
+        staticFiles("/", apiService.versionedStaticFilesPath)
+    }
+}
+
+private fun Route.installAboutRouteDocumentation() = installNotarizedRoute {
+    tags = setOf("API")
+
+    get = GetInfo.builder {
+        description("Get information about the API")
+        summary("Get about")
+        response {
+            description("Information about the API")
+            mediaTypes("application/json")
+            responseCode(HttpStatusCode.OK)
+            responseType<APIAbout>()
         }
     }
 }