+++ src/main/kotlin/app/revanced/api/configuration/services/ApiService.kt
@@ -12,6 +12,9 @@ internal class ApiService(
     private val backendRepository: BackendRepository,
     private val configurationRepository: ConfigurationRepository,
 ) {
+    val versionedStaticFilesPath = configurationRepository.versionedStaticFilesPath
+    val about = configurationRepository.about
+
     suspend fun contributors() = withContext(Dispatchers.IO) {
         configurationRepository.contributorsRepositoryNames.map {
             async {